Convert beaker provisioner total attempts to an integer data type (#121)
Converts the total attempts to wait for a beaker job to
finish from a float to an integer. It makes the logging look a bit
cleaner.

+++ teflo/provisioners/ext/bkr_client_plugin/beaker_client_plugin.py
@@ -224,7 +224,7 @@ class BeakerClientProvisionerPlugin(ProvisionerPlugin):
         self.logger.debug('Beaker timeout limit: %s.' % wait)
 
         # check Beaker status every 60 seconds
-        total_attempts = wait / 60
+        total_attempts = int(wait / 60)
 
         attempt = 0
         while wait > 0:
